+++ src/manifolds/GeneralUnitaryMatrices.jl
@@ -41,6 +41,33 @@ function GeneralUnitaryMatrices(
     return GeneralUnitaryMatrices{field, typeof(size), matrix_type}(size)
 end
 
+
+@doc raw"""
+    change_metric(M::GeneralUnitaryMatrices, ::EuclideanMetric, p, X)
+
+Change `X` to the corresponding vector with respect to the metric of the [`GeneralUnitaryMatrices`](@ref) `M`,
+which is just the identity, since the manifold is isometrically embedded.
+"""
+change_metric(M::GeneralUnitaryMatrices, ::EuclideanMetric, ::Any, ::Any)
+
+function change_metric!(::GeneralUnitaryMatrices, Y, ::EuclideanMetric, p, X)
+    copyto!(Y, X)
+    return Y
+end
+
+@doc raw"""
+    change_representer(M::GeneralUnitaryMatrices, ::EuclideanMetric, p, X)
+
+Change `X` to the corresponding representer of a cotangent vector at `p`.
+Since the [`GeneralUnitaryMatrices`](@ref) manifold `M` is isometrically embedded, this is the identity.
+"""
+change_representer(::GeneralUnitaryMatrices, ::EuclideanMetric, ::Any, ::Any)
+
+function change_representer!(::GeneralUnitaryMatrices, Y, ::EuclideanMetric, p, X)
+    copyto!(Y, X)
+    return Y
+end
+
 @doc raw"""
     check_point(M::UnitaryMatrices, p; kwargs...)
     check_point(M::OrthogonalMatrices, p; kwargs...)